recompiler: support for 64-bit enabled REM library on 32-bit hosts (disabled by USE_VBOXREM64 in Makefile.kmk now)
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@15272 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/recompiler_new/VBoxRecompiler.c
@@ -3492,7 +3492,7 @@ bool remR3DisasBlock(CPUState *env, int f32BitCode, int nrInstructions, char *ps
     /*
      * Disassemble.
      */
-    off = env->eip - (RTGCUINTPTR)pvPC;
+    off = env->eip - (RTGCUINTPTR)(uintptr_t)pvPC;
     Cpu.mode = f32BitCode ? CPUMODE_32BIT : CPUMODE_16BIT;
     Cpu.pfnReadBytes = NULL;            /** @todo make cs:eip reader for the disassembler. */
     //Cpu.dwUserData[0] = (uintptr_t)pVM;
@@ -3593,7 +3593,7 @@ bool remR3DisasInstr(CPUState *env, int f32BitCode, char *pszPrefix)
     /*
      * Disassemble.
      */
-    off = env->eip - (RTGCUINTPTR)pvPC;
+    off = env->eip - (RTGCUINTPTR)(uintptr_t)pvPC;
     Cpu.mode = f32BitCode ? CPUMODE_32BIT : CPUMODE_16BIT;
     Cpu.pfnReadBytes = NULL;            /** @todo make cs:eip reader for the disassembler. */
     //Cpu.dwUserData[0] = (uintptr_t)pVM;
